requirements.txt
gino~=1.0.1
asyncpg~=0.24.0
aiogram<3.0
aiohttp==3.6.2
aioredis~=2.0.0
python-dotenv~=0.19.0
SQLAlchemy~=1.3.24
requests~=2.26.0